core/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
dotenv==0.9.9
Faker==36.1.1
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
psycopg2-binary==2.9.9
djangorestframework_simplejwt==5.5.0
